Moved System tests to UserInterface tests
